Fix for Bug 121.
Issue the subscribe request against the origin not the provider.
We sometimes get situations where the subscribe request goes
and gets issues on the provider and reaches the origin before
it reaches the receiver.

The origin generates the ENABLE_SUBSCRIPTION as being originated
on the origin.  The ENABLE_SUBSCRIPTION can then reach the
receiver before the SUBSCRIBE_SET reaches the receiver.
This causes some of these node -1 errors.