Additional attempt at bug #40.
If libpq isn't already in the system shared library search
path directories then it might not be found when configure
tries to run the test program.

Build the path to libpq into the test executable via linker
flags.  The linker flags for this vary based on platform.
The logic was based on logic in postgresql/src/makefiles
in the Postgresql source tree.