Fixing issues when an admin conninfo does not have slony installed.
Querying the slony tables then checking if slony is installed
after the query fails was not a good idea because the failed
query would abort the transaction. This also creates
query failures in your postgresql log.  It is better
to first check if slony is installed then use the slony tables.
Created a slonik_is_slony_installed function to do this.

Also fixed some memory leaks in the get next id functions